refactor: separa logica de viewer login
